1402.03.07

+++ lib/Models/TempOrder.dart
@@ -2,7 +2,7 @@ class TempOrder{
 
   late String Name;
   late double Price;
-  late int SumAll;
+  late double SumAll;
   late int Count;
   late String Coka;
